+++ application/modules/master/views/rekening_index.php
@@ -47,7 +47,7 @@ function init_datatable() {
         'ajax': '<?php echo site_url('/master/rekening/datatable'); ?>',
         'stateSave': true,
         'order': [
-            [1, 'asc']
+            [2, 'asc']
         ],
         'fixedHeader': true,
         'columns': [{
@@ -57,7 +57,7 @@ function init_datatable() {
             {
                 orderable: false,
                 render: function(data, type, row, meta) {
-                    return '<a href="' + site_url + 'master/rekening/hapus/' + row.id +
+                    return '<a onclick="return confirm(\'Yakin untuk menghapus?\');" href="' + site_url + 'master/rekening/hapus/' + row.id +
                         '"><img src="<?php echo base_url(); ?>assets/img/del.png"></a>&nbsp;'+
                         '<a href="' + site_url + 'master/rekening/is_rekening_faktur/' + row.id +
                         '" class="btn btn-primary btn-circle btn-circle"><i class="ti-thumb-up"></i></a>';
